+++ extensions/chroma/lightIDgenerate.ts
@@ -163,6 +163,7 @@ export const LightIDList: {
    Panic2Environment: {},
    QueenEnvironment: {},
    LinkinPark2Environment: {},
+   TheRollingStonesEnvironment: {},
    GlassDesertEnvironment: {
       0: Array.from(Array(6), (_, i) => i + 1),
       1: Array.from(Array(8), (_, i) => i + 1),